Add a resource reference generator
Implements RFD 130.

See the included README for information about how the generator works.

This implementation breaks from RFD 130 in the following ways:

- The reference includes separate sections for dynamic resources and
  fields in order to make it easier to read.
- The generator specifies the kind and version of each dynamic resource.
  To do this, the generator looks up a method called for each resource
  type to assign the type's version and kind.
- The RFD was not clear about the possibility of overriding specific
  fields within a struct. This implementation provides this ability for
  situations where it is not possible to override the field's type
  definition, e.g., with standard library types like `time.Duration`.
- The generator uses a YAML configuration file. This is because the
  config ended up including more fields than specified in the RFD, so
  separating the config from the source made sense to give users less
  noise to deal with.

This change also edits comments within Go source files to add example
YAML overrides, and includes a generated reference.

Finally, this change adds tests for tools in `build.assets/tooling` to
the `test-go` make target so we can ensure that the resource reference
generator works and other tools in this directory work as expected.